+++ src/line_segment.rs
@@ -11,34 +11,34 @@ pub enum SolidSide {
 
 #[derive(Debug, Clone, Copy)]
 pub struct LineSegment {
-    pub start: Vector2<f32>,
-    pub end: Vector2<f32>,
+    pub start: Vector2<f64>,
+    pub end: Vector2<f64>,
     pub solid_side: SolidSide,
 }
 
 impl LineSegment {
-    pub fn new_both_solid(start: Vector2<f32>, end: Vector2<f32>) -> Self {
+    pub fn new_both_solid(start: Vector2<f64>, end: Vector2<f64>) -> Self {
         Self {
             start,
             end,
             solid_side: SolidSide::Both,
         }
     }
-    pub fn new_left_solid(start: Vector2<f32>, end: Vector2<f32>) -> Self {
+    pub fn new_left_solid(start: Vector2<f64>, end: Vector2<f64>) -> Self {
         Self {
             start,
             end,
             solid_side: SolidSide::Left,
         }
     }
-    pub fn add_vector(&self, vector: Vector2<f32>) -> Self {
+    pub fn add_vector(&self, vector: Vector2<f64>) -> Self {
         Self {
             start: self.start + vector,
             end: self.end + vector,
             solid_side: self.solid_side,
         }
     }
-    pub fn vector(&self) -> Vector2<f32> {
+    pub fn vector(&self) -> Vector2<f64> {
         self.end - self.start
     }
     fn left_solid_edge(&self) -> LeftSolidEdge {
@@ -50,7 +50,7 @@ impl LineSegment {
 }
 
 impl Collide for LineSegment {
-    fn aabb(&self, top_left: Vector2<f32>) -> Aabb {
+    fn aabb(&self, top_left: Vector2<f64>) -> Aabb {
         let start = self.start + top_left;
         let end = self.end + top_left;
         let x_min = start.x.min(end.x);
@@ -63,14 +63,14 @@ impl Collide for LineSegment {
     }
     fn for_each_edge_facing<F: FnMut(LineSegment)>(
         &self,
-        _direction: Vector2<f32>,
+        _direction: Vector2<f64>,
         mut f: F,
     ) {
         f(*self);
     }
-    fn for_each_vertex_facing<F: FnMut(Vector2<f32>)>(
+    fn for_each_vertex_facing<F: FnMut(Vector2<f64>)>(
         &self,
-        direction: Vector2<f32>,
+        direction: Vector2<f64>,
         mut f: F,
     ) {
         self.for_each_left_solid_edge_facing(direction, |edge| {
@@ -81,7 +81,7 @@ impl Collide for LineSegment {
 
     fn for_each_left_solid_edge_facing<F: FnMut(LeftSolidEdge)>(
         &self,
-        direction: Vector2<f32>,
+        direction: Vector2<f64>,
         mut f: F,
     ) {
         let vector = self.vector();
