+++ src/shape.rs
@@ -13,7 +13,7 @@ pub enum Shape {
 }
 
 impl Shape {
-    pub fn aabb(&self, top_left: Vector2<f32>) -> Aabb {
+    pub fn aabb(&self, top_left: Vector2<f64>) -> Aabb {
         match self {
             &Shape::AxisAlignedRect(ref rect) => rect.aabb(top_left),
             &Shape::LineSegment(ref line_segment) => line_segment.aabb(top_left),
@@ -21,10 +21,10 @@ impl Shape {
     }
     pub fn movement_collision_test(
         &self,
-        position: Vector2<f32>,
+        position: Vector2<f64>,
         stationary: &Self,
-        stationary_position: Vector2<f32>,
-        movement_vector: Vector2<f32>,
+        stationary_position: Vector2<f64>,
+        movement_vector: Vector2<f64>,
     ) -> MovementWithSlide {
         match self {
             &Shape::AxisAlignedRect(ref moving) => match stationary {
